pref: move onload outside.
